border-radius & pricing section
I have an issue with the border-radius, I wasn't able to apply it on the container that I wanted.  Workaround was to put border-radius on the corners of the top and bottom cards.

Also an issue was with the middle card that contains the "$29 per month": I had some difficulty aliging it correctly, while trying to use flex-box. I used the dev tool in chrome to find that the span was using the left space, which prevented justify-content: flex-start; to work.  Amended span to have margin-left: 0; which corrected it.