Attempting media query
Since I have tried to do mobile-first, I'm now trying to get the layout correct using @media for the desktop design.
I've also realised that my initial sizing is possibly too big and need to investigate how to stop it running off screen. I believe the error was that I was not at correct size when viewing with dev tools!